mitosheet: move out of preprocessing steps
