+++ mitosheet/mitosheet/preprocessing/preprocess_step_performer.py
@@ -31,7 +31,7 @@ class PreprocessStepPerformer(ABC, object):
 
     @classmethod
     @abstractmethod
-    def execute(cls, args: Collection[Any], kwargs: Dict[str, Any]) -> Tuple[List[Any], Optional[List[str]], Optional[Dict[str, Any]]]:
+    def execute(cls, args: Collection[Any]) -> Tuple[List[Any], Optional[List[str]], Optional[Dict[str, Any]]]:
         """
         Execute always returns the new list of arguments, as well as the names of these arguements, as well as execution_data
         for this preprocess step.
